we are going with slider for now
